refactor: become the repo to a monorepo
