+++ .gitignore
@@ -1,6 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
+apps/*/node_modules
 /node_modules
 /.pnp
 .pnp.js
@@ -13,9 +14,13 @@
 /prisma/db.sqlite-journal
 
 # next.js
+apps/*/.next/
 /.next/
 /out/
 next-env.d.ts
+
+# turbo
+.turbo
 
 # production
 /build
@@ -32,6 +37,7 @@ yarn-error.log*
 
 # local env files
 # do not commit any .env files to git, except for the .env.example file. https://create.t3.gg/en/usage/env-variables#using-environment-variables
+apps/*/.env
 .env
 .env*.local
