+++ .vscode/settings.json
@@ -1,0 +1,15 @@
+{
+	"editor.tabSize": 2,
+	"editor.codeActionsOnSave": {
+		"source.fixAll.eslint": true
+	},
+	"editor.formatOnSave": true,
+	"todohighlight.keywords": [
+		{
+			"text": "NOTE:",
+			"color": "#FFFFFF",
+			"backgroundColor": "#6D87A4",
+			"overviewRulerColor": "#FFFFFF"
+		}
+	]
+}
