+++ apps/bot/package.json
@@ -1,0 +1,21 @@
+{
+	"name": "bot",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"start": "ts-node src/index.ts",
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"discord.js": "^14.7.1"
+	},
+	"devDependencies": {
+		"@types/node": "^18.11.18",
+		"ts-node": "^10.9.1",
+		"typescript": "^4.9.4"
+	}
+}
