+++ apps/bot/src/commands/_commandList.ts
