+++ apps/bot/src/commands/wallet.ts
