+++ apps/bot/src/config.ts
@@ -2,11 +2,12 @@ import { GatewayIntentBits } from "discord.js";
 // import { env } from "@/env.mjs";
 
 export const config = {
-  prefix: "!",
-  token: "",
+	prefix: "!",
+	token: "",
 };
 
 export const intentOptions: Array<GatewayIntentBits> = [
-  GatewayIntentBits.Guilds,
-  GatewayIntentBits.GuildMessages,
+	GatewayIntentBits.Guilds,
+	GatewayIntentBits.GuildMessages,
+	GatewayIntentBits.MessageContent,
 ];
