+++ apps/bot/src/types/command.ts
