+++ apps/bot/src/types/discord.d.ts
