+++ apps/bot/tsconfig.json
@@ -1,0 +1,24 @@
+{
+	"compilerOptions": {
+		"target": "ESNext",
+		"module": "commonjs",
+		"rootDir": "./src/",
+		"outDir": "./dist/",
+		"strict": true,
+		"moduleResolution": "node",
+		"importHelpers": true,
+		"experimentalDecorators": true,
+		"esModuleInterop": true,
+		"skipLibCheck": true,
+		"allowSyntheticDefaultImports": true,
+		"resolveJsonModule": true,
+		"forceConsistentCasingInFileNames": true,
+		"removeComments": true,
+		"typeRoots": ["node_modules/@types"],
+		"sourceMap": false,
+		"baseUrl": "./"
+	},
+	"files": ["src/index.ts"],
+	"include": ["./**/*.ts"],
+	"exclude": ["node_modules", "dist"]
+}
