+++ apps/main/.env.example
