+++ apps/main/next.config.mjs
