+++ apps/main/public/favicon.ico
