+++ apps/main/src/pages/api/auth/[...nextauth].ts
