+++ apps/main/src/styles/globals.css
