+++ apps/main/src/utils/get-issue-description.js
