+++ apps/main/src/utils/send-discord-message.js
