+++ package.json
@@ -1,53 +1,19 @@
 {
-  "name": "reward-system",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "next build",
-    "dev": "next dev",
-    "postinstall": "prisma generate",
-    "lint": "next lint",
-    "start": "next start"
-  },
-  "dependencies": {
-    "@actions/core": "^1.10.0",
-    "@next-auth/prisma-adapter": "^1.0.5",
-    "@octokit/rest": "^19.0.7",
-    "@prisma/client": "^4.9.0",
-    "@tanstack/react-query": "^4.20.2",
-    "@trpc/client": "^10.9.0",
-    "@trpc/next": "^10.9.0",
-    "@trpc/react-query": "^10.9.0",
-    "@trpc/server": "^10.9.0",
-    "discord.js": "^14.7.1",
-    "next": "13.1.6",
-    "next-auth": "^4.19.0",
-    "node-fetch": "^3.3.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "superjson": "1.9.1",
-    "zod": "^3.20.2"
-  },
-  "devDependencies": {
-    "@types/eslint": "^8.21.1",
-    "@types/node": "^18.11.18",
-    "@types/prettier": "^2.7.2",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.47.1",
-    "@typescript-eslint/parser": "^5.47.1",
-    "autoprefixer": "^10.4.7",
-    "eslint": "^8.30.0",
-    "eslint-config-next": "13.1.6",
-    "postcss": "^8.4.14",
-    "prettier": "^2.8.1",
-    "prettier-plugin-tailwindcss": "^0.2.1",
-    "prisma": "^4.9.0",
-    "tailwindcss": "^3.2.0",
-    "typescript": "^4.9.4"
-  },
-  "ct3aMetadata": {
-    "initVersion": "7.5.6"
-  }
+	"private": true,
+	"workspaces": [
+		"apps/*",
+		"packages/*"
+	],
+	"scripts": {
+		"build": "turbo run build",
+		"dev": "turbo run dev --parallel",
+		"clean": "find . -name node_modules -o -name .next -o -name .expo -type d -prune | xargs rm -rf",
+		"lint": "turbo run lint",
+		"start": "turbo run start",
+		"format": "prettier --write \"**/*.{ts,tsx,md}\""
+	},
+	"devDependencies": {
+		"prettier": "latest",
+		"turbo": "latest"
+	}
 }
